requirements.txt
# Core dependencies
python-telegram-bot==20.1
python-dotenv==1.1.0

# HTTP and networking
httpx>=0.23.3
httpcore>=0.16.3
h11>=0.14.0

# Async support
anyio>=4.9.0
sniffio>=1.3.1

# Utilities
certifi>=2025.4.26
idna>=3.10
typing_extensions>=4.13.2

# Development dependencies (optional)
# pytest==7.4.0
# black==23.3.0
# flake8==6.0.0
